backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
librosa==0.10.1
openai==1.3.7
pydub==0.25.1
python-multipart==0.0.6 